Resolves IOOperator into ByteIOOperator and CharIOOperator
